% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenMapAPI.R
\name{getData}
\alias{getData}
\title{run the https request to get the data}
\arguments{
\item{year}{parameter description}

\item{dataSource}{parameter description}
}
\description{
method description. method description
}
