% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenMapAPI.R
\name{getSource}
\alias{getSource}
\title{get source according to timeline}
\arguments{
\item{year}{parameter description}

\item{dataSource}{parameter description}
}
\description{
method description. method description
}
