\name{Lab05R-package}
\alias{Lab05R-package}
\alias{Lab05R}
\docType{package}
\title{
\packageTitle{Lab05R}
}
\description{
\packageDescription{Lab05R}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab05R}
\packageIndices{Lab05R}
~~ An overview of how to use the package, including the most important functions ~~
}

