% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenMapAPI.R
\name{multiData}
\alias{multiData}
\title{get data by year and datasoruce}
\arguments{
\item{year}{parameter description}

\item{dataSource}{parameter description}
}
\description{
method description. method description
}
