% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenMapAPI.R
\docType{class}
\name{thenmapsApi}
\alias{thenmapsApi}
\title{thenmap API}
\description{
Calculate greatest common divisor
The class for linear regression
}
\section{Fields}{

\describe{
\item{\code{center_cor}}{list of lat and lbg.}
}}


